requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.39.0
bz2file==0.98
celery==3.1.23
Django==1.9.4
django-cors-headers==1.1.0
djangorestframework==3.3.3
djoser==0.4.3
drf-extensions==0.2.8
gunicorn==19.4.5
httpretty==0.8.10
jsonrpclib==0.1.7
kombu==3.0.34
nltk==3.2
psycopg2==2.6.1
pytz==2016.1
redis==2.10.5
requests==2.8.1
six==1.10.0
smart-open==1.3.2
wheel==0.24.0
factory-boy==2.7.0